+++ src/__tests__/runtime.ts
@@ -1,11 +1,20 @@
-import {Expressions, Snapshots} from '../definitions';
 import {Runtime} from '../runtime';
 
 let runtime: Runtime;
 beforeAll(() => {
-  const snapshots: Snapshots = {0: 'snapshot'};
-  const expressions: Expressions = {0: 'expression'};
-  runtime = new Runtime(expressions, snapshots);
+  runtime = new Runtime({
+    0: {
+      line: 0,
+      snapshot: 'snapshot 0',
+      expression: 'expression 0',
+    },
+    1: {
+      line: 1,
+      snapshot: 'snapshot 1',
+      expression: 'expression 1',
+      description: 'description 1',
+    },
+  });
 });
 
 test('#snapshot() should return correctly', () => {
@@ -14,4 +23,8 @@ test('#snapshot() should return correctly', () => {
 
 test('#report() should return correctly', () => {
   expect(runtime.report(0)).toMatchSnapshot();
+});
+
+test('#report() should return correctly with description', () => {
+  expect(runtime.report(1)).toMatchSnapshot();
 });