+++ src/utils/create-setup-expression.ts
@@ -1,17 +1,15 @@
-import {config_namespace, package_name, runtime_namespace, Expressions, Trigger} from '../definitions';
+import {config_namespace, package_name, runtime_namespace, Target, Trigger} from '../definitions';
 import {remove_spaces} from './remove-spaces';
 
 export const create_setup_expression = (triggers: Trigger[]) => {
-  const expressions = triggers.reduce<Expressions>(
-    (infos, trigger) => ({
-      ...infos,
-      [trigger.line]: trigger.expression,
-    }),
-    {},
+  const targets = triggers.reduce<Target[]>(
+    (current_targets, {line, description, expression}) =>
+      ([...current_targets, {line, description, expression}]),
+    [],
   );
 
   const stringified_package_name = JSON.stringify(package_name);
-  const stringified_expressions = JSON.stringify(expressions);
+  const stringified_targets = JSON.stringify(targets);
 
   const config_expression = remove_spaces(`
     (function () {
@@ -27,7 +25,7 @@ export const create_setup_expression = (triggers: Trigger[]) => {
     var ${runtime_namespace};
     beforeAll(function () {
       ${runtime_namespace} = require(${stringified_package_name})
-        .setup(module.filename, ${config_expression}, ${stringified_expressions});
-    });
+        .setup(module.filename, ${config_expression}, ${stringified_targets});
+    })
   `);
 };